Common source code merge
- Update poll timer callback's to use 64bit integers for delay.
- Avoid creating substrs if unnecessary in Unicode_CompareRange.
- Vmxnet3: Add Receive Data Ring Support.
- Comment cleanup.

+++ open-vm-tools/lib/include/poll.h
@@ -307,7 +307,7 @@ Bool Poll_CB_DeviceRemove(PollerFunction f,
 
 VMwareStatus Poll_CB_RTime(PollerFunction f,
                            void *clientData,
-                           int delay,	// microseconds
+                           int64 delay,   // microseconds
                            Bool periodic,
                            MXUserRecLock *lock);
